Миграция на версию ImGui 1.83 Миграция на версию glm 0.9.9.8 Исправлена несовместимость imgui_dock с ImGui 1.83 Рефакторинг imgui_dock Исправлена ошибка в imgui_curve Исправлена регистрация update в ParticleViewer

+++ include/Utils/Curve.h
@@ -6,8 +6,8 @@
 
 namespace Curve {
     void Spline(const float *key, int num, int dim, float t, float *v);
-    f32 CurveValue(f32 position, u8 maxPoints, const glm::vec2 *points);
-    f32 CurveValueSmooth(f32 position, i32 maxPoints, const glm::vec2 *points);
+    f32 CurveValue(f32 position, u32 maxPoints, const glm::vec2 *points);
+    f32 CurveValueSmooth(f32 position, u32 maxPoints, const glm::vec2 *points);
 };
 
 #endif
